+++ test/fixtures/controllers/comments/show.js
@@ -1,0 +1,15 @@
+"use strict";
+
+import Controller from "../../../../src/controller";
+
+export default class ShowComment extends Controller {
+  model(req, res, next) {
+    res.send({
+      foo: {
+        id: 1,
+        name: "bar"
+      }
+    });
+    next();
+  }
+}